+++ src/app/features/currencies/models/create-currency-payload.ts
@@ -1,4 +1,4 @@
-export interface CurrencyCreationPayload {
+export interface CreateCurrencyPayload {
 	name: string; // 1-64
 	alphabetical_code?: string; // 3, alphabetical code
 }