+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 - Added currencies page with paginated currencies table.
 - Added currency creation modal.
 - Added currency update modal.
+- Added currency delete confirm.
+- Added currencies delete all confirm.
+- Added deletion of all expenses that use a currency.
 ### Changed
 - For changes in existing functionality.
 ### Deprecated